refactoring, fixes
